+++ libp2p/CHANGELOG.md
@@ -1,4 +1,9 @@
-## 0.55.1
+## 0.56.0
+
+- Remove deprecated `Transport::with_bandwidth_logging`,
+  `SwarmBuilder::with_bandwidth_logging` and `TransportExt`. 
+  See [PR 5766](https://github.com/libp2p/rust-libp2p/pull/5766).
+
 - Introduce `libp2p-webrtc-websys` behind `webrtc-websys` feature flag.
   See [PR 5819](https://github.com/libp2p/rust-libp2p/pull/5819).